requirements_test.txt
# for manual testing only
python-decouple